% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing_temporal_plot.R
\name{forcingvar}
\alias{forcingvar}
\title{Plot SUMMA input variable for a given HRU}
\usage{
forcingvar(ncname, ncvar, hruindex)
}
\arguments{
\item{ncname}{Filename (with filepath) of the SUMMA forcing netCDF file produced by MetSim}

\item{hruindex}{HRU Index of interest in SUMMA forcing}

\item{varname}{Variable of interest in SUMMA forcing}
}
\description{
Variables available for plotting \cr \cr
pptrate = Precipitation Rate \cr
SWRadAtm = Shortwave Radiation \cr
LWRadAtm = Longwave Radiation \cr
windspd = Wind Speed \cr
spechum = Specific Humidity \cr
airpres = Air Pressure \cr
}
\examples{
forcingvar('data_raw/forcing/forecastforcing_modified.nc','pptrate', 1)
}
