% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_HRU.R
\name{nearestHRU}
\alias{nearestHRU}
\title{Geographically closest HRU to given coordinates}
\usage{
nearestHRU(ncname, uLat, uLong)
}
\arguments{
\item{ncname}{Filename (with filepath) of the SUMMA Local Attributes netCDF file}

\item{uLat}{Latitude}

\item{uLong}{Longitude}
}
\value{
HRU Index that is nearest to uLat and uLong
}
\description{
This function finds the HRU Index that is closest to a given latitude and longitude
}
\examples{
nearestHRU('data_raw/summa_zLocalAttributes_columbia_gru.nc', 47.277603, -121.786773)
}
