% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job_list.R
\name{summa}
\alias{summa}
\title{Parallelization of SUMMA run}
\usage{
summa(config, nGRU)
}
\arguments{
\item{config}{SUMMA Confg file}

\item{nGRU}{Length of each GRU set for a SUMMA run}
}
\description{
Parallelization of SUMMA run
}
\examples{
summa('/d3/msaharia/SHARP/summa/PNW2/summa_fileManager_PNW.txt', 20)
}
