% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_temporaloutputdata.R
\name{temporaloutputdata}
\alias{temporaloutputdata}
\title{Extracts variable data from a SUMMA output file for a single HRU}
\usage{
temporaloutputdata(ncfile, hruIndex)
}
\arguments{
\item{ncfile}{#SUMMA output file in netCDF format}

\item{hruIndex}{#Index of the HRU set for data is sought}
}
\value{
Dataframe of variables contained in the SUMMA output netCDF file
}
\description{
Extracts variable data from a SUMMA output file for a single HRU
}
\examples{
sumdata_all <- temporaloutputdata('vault/benchmarking/PNW_3L_1H.nc', 1)
}
