% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zDecisions_files.R
\name{zDecs}
\alias{zDecs}
\title{Create zDecision files for every month automatically}
\usage{
zDecs(zDecFile, startDate, nMonths)
}
\arguments{
\item{zDecFile}{zDecisions template file}

\item{startDate}{Start date of the simulation}

\item{nMonths}{Number of months from the start date for which files are to be created}
}
\description{
Create zDecision files for every month automatically
}
\examples{
zDecs('data_raw/PNW2/summa_zDecisions_template.txt', '2016-10-01 00:00', 20)
}
